requirements.txt
gspread
oauth2client
pandas
numpy
spacy
scikit-learn
seaborn
matplotlib
plotly
sentence-transformers
faiss-cpu
fastapi
uvicorn
requests
apscheduler
streamlit
groq
load_dotenv
scikit-learn